refactor(apiclients): split client.go into wrapped_client/doer/error
Break the single hand-written client.go into focused files:
- wrapped_client.go: endpoint/timeout/retry defaults, WrappedClient, New
- doer.go: authDoer (+ a HttpRequestDoer compile-time assertion)
- error.go: APIError, ErrorFromResponse, decodeError

No behavior change.
